+++ RELEASENOTES.md
@@ -6,6 +6,8 @@
 
 **BREAKING CHANGES**
 - Changes to how the `inverse` prop works for icons and removal of the default value of `true`
+- Button component: Inverse style of Button (white color to go on dark backgrounds) is now declared with a bool prop, inverse
+- Button component: `inverse` and `icon-inverse` are no longer supported in the variants prop
 
 **MAJOR CHANGES**
 - Add support for custom svg icons
@@ -13,6 +15,8 @@
 
 **OTHER**
 - Allow menus and buttons used in the Global Nav Bar to be marked as active
+- Button component: when button icon had an iconVariant prop, the iconSize prop did not render. It is now fixed
+- Update the logo used in the Global Header story
 
 
 ## Release 0.1.2
@@ -38,7 +42,6 @@
 - Add Button support to Global Navigation Bar
 - Update the keyboard navigation logic used by menus
 - Update profile icon used in the Global Header
-
 
 ## Release 0.1.0